chore: Decluttered express routes (#447)
* Deformatted express routes, fix invites not being sent

* Added healthcheck route back

* Removed method not allowed

* Delete org route with org id

* Updated next to latest

+++ components/Dashboard/DashboardContent.tsx
@@ -75,7 +75,7 @@ export default function DashboardContent() {
     }
 
     try {
-      const { message } = await OrgsService.deleteOrg();
+      const { message } = await OrgsService.deleteOrg(user?.orgId);
       alert(message);
     } catch (error) {
       alert(error.response.data.message);
